# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Office Works",
    "description": "Production-grade FastAPI backend with auth (JWT+OAuth+RBAC) and LLM chat proxy domains.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/chat/complete": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Non-streaming chat completion",
        "description": "Send a conversation to the active LLM provider and receive the full response in a single JSON body.  Provider is set via ``LLM_PROVIDER``.",
        "operationId": "chat_complete_api_v1_chat_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Streaming chat completion (SSE)",
        "description": "Stream the LLM response as Server-Sent Events.  Each ``data`` event carries a text chunk.  A final ``data: [DONE]`` event signals the end.",
        "operationId": "chat_stream_api_v1_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/chat/conversations": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List conversations for authenticated user",
        "description": "Return all conversations for the authenticated user (newest first).",
        "operationId": "list_conversations_api_v1_chat_conversations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationResponse"
                  },
                  "type": "array",
                  "title": "Response List Conversations Api V1 Chat Conversations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create a new conversation",
        "description": "Create a new conversation session for the authenticated user.\n\nRequires the ``chat:write`` permission.",
        "operationId": "create_conversation_api_v1_chat_conversations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chat/conversations/{conversation_id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get a conversation",
        "description": "Return a conversation owned by the authenticated user.",
        "operationId": "get_conversation_api_v1_chat_conversations__conversation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/conversations/{conversation_id}/messages": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List messages in a conversation",
        "description": "Return all messages in a conversation (oldest first).",
        "operationId": "list_messages_api_v1_chat_conversations__conversation_id__messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response List Messages Api V1 Chat Conversations  Conversation Id  Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send a message — SSE streaming + DB persistence",
        "description": "Send a user message and receive an SSE-streamed assistant reply.\n\nWorkflow\n--------\n1. Verify conversation ownership.\n2. Persist the user message.\n3. Rebuild LangChain message list from conversation history.\n4. Stream tokens from the LLM provider.\n5. Collect all chunks; persist the completed assistant message.\n6. If this is the first turn, auto-generate a title.\n\nReturns an :class:`~sse_starlette.sse.EventSourceResponse` when\n``body.stream=True`` (default); returns a :class:`MessageResponse` JSON\nbody when ``body.stream=False``.\n\nRequires the ``chat:write`` permission.",
        "operationId": "send_message_api_v1_chat_conversations__conversation_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Send Message Api V1 Chat Conversations  Conversation Id  Messages Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "title": "Messages"
          },
          "system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "ChatRequest",
        "description": "Request body for chat endpoints.\n\nAttributes\n----------\nmessages:\n    Ordered conversation history.  Must contain at least one message.\nsystem:\n    Optional system prompt prepended before all other messages.\n    Equivalent to adding a ``{\"role\": \"system\", ...}`` entry at the front."
      },
      "ChatResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatResponse",
        "description": "Response body for the non-streaming ``/complete`` endpoint.\n\nAttributes\n----------\ncontent:\n    The model's full reply text.\nmodel:\n    Active LiteLLM model identifier, e.g. ``\"openai/gpt-4o-mini\"``.\n    Included for debugging and provider-routing verification."
      },
...
      "ConversationCreate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          }
        },
        "type": "object",
        "title": "ConversationCreate",
        "description": "Request body for POST /chat/conversations."
      },
      "ConversationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "system_prompt",
          "model_name",
          "created_at",
          "updated_at"
        ],
        "title": "ConversationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "token_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Count"
          },
          "finish_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "conversation_id",
          "role",
          "content",
          "token_count",
          "finish_reason",
          "created_at"
        ],
        "title": "MessageResponse"
      },
...
      "SendMessageRequest": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 32000,
            "minLength": 1,
            "title": "Content"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "description": "When True, response is SSE-streamed. When False, returns full JSON.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "SendMessageRequest",
        "description": "Request body for POST /chat/conversations/{id}/messages."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}